EFM - Whites / 4.39 rebalance
Some underpowered that is about to be fixed.

- Terror Rider HP to 37 (+1) , XP to 43 (-5)
- Terror Strider HP to 62 (+6) 
Who cares about more health, this unit is just quite underpowered.
- Warder HP to 37 (+1) , XP to 34 (-3) , price to 12g (-1g)
- Guardian HP to 55 (+5)
- Sentinal HP to 68 (+5)
There was a need to make this unit so terrible? another buff from another review...
- Scout HP to 31 (+1) , XP to 34 (-2)
- Raider ranged damage to 8-3 (+1)
It has no lv3 ...
- Bowlady price to 56g (+3g)
- Storm Witch price to 37g (+3g)
- Storm Lady XP to 150 (+16) , price to 66g (+4g)
The only nerf... but considering the buffs on the bad units might be a WIN for the faction.